🚚 rename [#97] Card api mock 파일의 경로 변경
